refactor(ale_claw): move image_sanitization into harness/utils/ leaf package
image_sanitization is a pure bytes->bytes transform (no internal imports),
shared by the loop, the read tool, and analyze_image. Homed in utils/ as a
leaf util (pi convention) rather than floating at base. Behavior-preserving:
git mv + import-path updates. 928 passed, 3 skipped.
