+++ ale_run/agents/ale_claw/harness/utils/image_sanitization.py
